Refactored the sequence table into its own component
